Move models to correct files
