+++ HausClock/Models/Game.swift
@@ -1,0 +1,15 @@
+//
+//  Game.swift
+//  HausClock
+//
+//  Created by Tom Brown on 11/28/14.
+//  Copyright (c) 2014 not. All rights reserved.
+//
+
+import Foundation
+
+enum GameState {
+    case Active
+    case Paused
+    case Finished
+}
